requirements/base.txt
Django==2.2.5
django-allauth==0.40.0
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-model-utils==3.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
psycopg2-binary==2.8.3
PyYAML==5.1.2
django-rest-swagger==2.2.0
Pillow==6.1.0

